+++ src-docs/src/views/accordion/accordion_multiple.js
@@ -19,7 +19,8 @@ export default () => (
 
     <EuiAccordion
       id="accordion2"
-      buttonContent="A second accordion with padding"
+      buttonContent="A second accordion with padding and a very long title that should truncate because of eui-textTruncate"
+      buttonContentClassName="eui-textTruncate"
       paddingSize="l">
       <EuiText>
         <p>The content inside can be of any height.</p>